// src/models/Store.ts
import {User} from './User';
import {Feedback} from './Feedback';
import {Product} from './Product';
import {Categorie} from './Categorie';
import {Voucher} from './Voucher';
export interface Store {
  "id": number;
  "user_id": number;
  "favorite": boolean;
  "name": string;
  "address": string;
  "postal_code": string;
  "district": string;
  "city": string;
  "lat": string;
  "long": string;
  "start_time": string;
  "end_time": string;
  "monday": number;
  "tuesday": number;
  "wednesday": number;
  "thursday": number;
  "friday": number;
  "saturday": number;
  "sunday": number;
  "created_at": string;
  "updated_at": string;
  "commission_rate": string;
  "distance": string;
  "avatar": string;
  "avatar_files": string;
  "user" : User | null;
  "averageStarRating": number;
  "description": string;
  "feedbacks" : Feedback | null;
  "popular_products": Product[] | null;
  "categories": Categorie[] | null;
  "vouchers" : Voucher[] | null;
}